SLID-Librosa/requirements.txt
librosa==0.10.1
matplotlib==3.8.2
pydub==0.25.1
loguru==0.7.2
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.3